scripts/requirements.txt
python
pyyaml
jinja2
